[JENKINS-32160] Set StrictHostKeyChecking=no even if no credentials are provided
This solves a confusing situation where the user configures a job with a
ssh repository, but doesn't provide ssh credentials.

In this case, the plugin doesn't provide StrictHostKeyChecking=no to the
GIT_SSH configuration, and the build likely fails with "Host key verification
failed." unless the build slave has been pre-configured with the host key.

Then, the user tries to fix this known hosts issue, rather than providing
a ssh key which is the real fix.